gata

+++ app/src/main/java/com/example/narcis/zvonnecontrol/obiecte/coman.java
@@ -9,11 +9,11 @@ public class coman {
     String nume;
     String data;
     String nrdetelefon;
-    int id;
+    long id;
     int status;
 
     public coman(){}
-    public coman(String text, String nume, String data, int status, int id,String numar) {
+    public coman(String text, String nume, String data, int status, long id,String numar) {
         this.text = text;
         this.nume = nume;
         this.data = data;
@@ -62,11 +62,11 @@ public class coman {
         this.status = status;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 }
